Move API custom types into shared locations (updating where required) allowing removal of App versions
