+++ thoth-api/src/graphql/mod.rs
@@ -1,1 +1,3 @@
+#[cfg(feature = "backend")]
 pub mod model;
+pub mod utils;
